Refactor app bootstrap into dedicated modules
